backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-cors-headers
psycopg2-binary
gunicorn
pytest-django
dj-database-url
whitenoise
